Adding the transaction status and order second trial

+++ src/main/java/innovitics/azimut/rest/mappers/GetTransactionsMapper.java
@@ -90,7 +90,7 @@ public class GetTransactionsMapper extends RestMapper<GetTransactionsInput,GetTr
 			
 			if(transactionOutput.getStatusType()!=null)
 			{
-				businessTransaction.setStatus(TransactionStatus.getById(transactionOutput.getStatusType()).getStatusId());				
+				businessTransaction.setStatus(TransactionStatus.getById(transactionOutput.getStatusType().intValue()).getStatusId());				
 			}
 
 			if(transactionOutput.getOrderType()!=null)
